requirements/base.txt
Django==2.0.7
djangorestframework==3.7.7
pytz==2018.5
django-environ==0.4.4
djangorestframework-jwt==1.11.0
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.6.1
six==1.11.0
sqlparse==0.2.4
django-rest-swagger==2.1.2
